Fixes #3. Black and white squares can now be indicated.
This commit adds functions to set a tile as either black (no letter) or
white (definitely contains a letter). The new statuses are represented
with new css classes.

This commit also updates the styling of the page slightly, adding in a
background image.

Double-clicking an empty tile will cycle the tile through three states,
in this order: black -> white -> empty -> black.

If a letter is entered by the user into any tile, that tile will
automatically be set as a white tile.