(Fixes #1) Allow other configurations than 3x3.
This commit allows the page to accept Junimojis with a different number
of rows and columns than 3 by 3. The grid will automatically resize and
adapt to display the correct layout.

The number of rows and colums is specified by two new URL query parameters:
  r=
  c=
So, for a 4x4 grid, the params would be "?r=4&c=4". This would produce a
4x4 grid, with 16 total sections.

The colour scheme is extended so that it repeats for sections beyond the
9th (index 8).